Move GenericTablePanel to another package
