+++ orienteer-core/src/main/java/org/orienteer/core/component/table/component/GenericTablePanel.java
@@ -1,4 +1,4 @@
-package org.orienteer.core.component.table.filter.component;
+package org.orienteer.core.component.table.component;
 
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.ajax.markup.html.form.AjaxFallbackButton;
