+++ orienteer-core/src/main/java/org/orienteer/core/component/table/component/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package contains filter for wicket data tables
+ */
+package org.orienteer.core.component.table.component;
