--- orienteer-core/src/main/java/org/orienteer/core/component/table/filter/component/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package contains filter for wicket data tables
- */
-package org.orienteer.core.component.table.filter.component;
