+++ orienteer-core/src/main/java/org/orienteer/core/component/widget/AbstractCalculatedDocumentsWidget.java
@@ -15,7 +15,7 @@ import org.orienteer.core.component.command.ExportCommand;
 import org.orienteer.core.component.command.SaveODocumentsCommand;
 import org.orienteer.core.component.property.DisplayMode;
 import org.orienteer.core.component.table.OrienteerDataTable;
-import org.orienteer.core.component.table.filter.component.GenericTablePanel;
+import org.orienteer.core.component.table.component.GenericTablePanel;
 import org.orienteer.core.service.impl.OClassIntrospector;
 import org.orienteer.core.widget.AbstractWidget;
 import ru.ydn.wicket.wicketorientdb.model.OQueryDataProvider;
