+++ orienteer-core/src/main/java/org/orienteer/core/component/widget/browse/ListAllODocumentsWidget.java
@@ -11,7 +11,7 @@ import org.orienteer.core.component.FAIconType;
 import org.orienteer.core.component.command.*;
 import org.orienteer.core.component.property.DisplayMode;
 import org.orienteer.core.component.table.OrienteerDataTable;
-import org.orienteer.core.component.table.filter.component.GenericTablePanel;
+import org.orienteer.core.component.table.component.GenericTablePanel;
 import org.orienteer.core.component.table.filter.sql.ODefaultQueryBuilder;
 import org.orienteer.core.service.IOClassIntrospector;
 import org.orienteer.core.widget.AbstractWidget;
